refactor(spa): split store.test.ts into workspace and tabs test files (#254)
Split 611-line store.test.ts into:
- store-workspace.test.ts (199 lines, 21 tests): CRUD, reorder, settings, import, moduleConfig
- store-tabs.test.ts (431 lines, 30 tests): tab ops, insertTab, closeTabInWorkspace

Total test count preserved: 21 + 30 = 51.
